requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
asyncio>=3.4.3

# Twitch integration
twitchAPI>=4.0.0

# OBS integration
obs-websocket-py>=1.0.0

# Discord integration
discord.py>=2.3.0

# Development tools
black>=23.0.0
pylint>=2.17.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
mypy>=1.3.0

# MCP dependencies
mcp>=0.3.0
pydantic-ai>=0.0.4
